deleting not necessary files
