+++ src/index.ts
@@ -1,5 +1,5 @@
 import ARToolkitNFT from './ARToolkitNFT'
-import ARControllerNFT from './ARControllerNFT2'
+import ARControllerNFT from './ARControllerNFT'
 
 export default {
   ARToolkitNFT, ARControllerNFT
