+++ types/src/index.d.ts
@@ -1,5 +1,5 @@
 import ARToolkitNFT from './ARToolkitNFT';
-import ARControllerNFT from './ARControllerNFT2';
+import ARControllerNFT from './ARControllerNFT';
 declare const _default: {
     ARToolkitNFT: typeof ARToolkitNFT;
     ARControllerNFT: typeof ARControllerNFT;
